+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [2.0.1] - 2018-05-10
 
 ## Fixed
 
@@ -43,5 +43,6 @@
 
 
 
-[Unreleased]: https://github.com/sdispater/pendulum/compare/2.0.0...master
+[Unreleased]: https://github.com/sdispater/pendulum/compare/2.0.1...master
+[2.0.1]: https://github.com/sdispater/pendulum/releases/tag/2.0.1
 [2.0.0]: https://github.com/sdispater/pendulum/releases/tag/2.0.0